Refactor: - Split into several files - Use ES6 classes
