+++ lib/Endianness.js
@@ -1,0 +1,9 @@
+"use strict";
+
+/**
+ * @enum {number}
+ */
+module.exports = {
+    BIG: 0,
+    LITTLE: 1
+};
